fix: seguridad — IDOR, CORS, auth y tipos de datos
- SecurityConfig: endpoints protegidos por JWT (solo /auth, /publico, /whatsapp son publicos)
- Todos los controllers: remover @CrossOrigin(origins="*"), CORS manejado por SecurityConfig
- FaenaController, GastoController, IngresoController, MantenimientoController, SalarioController, PagoClienteController, OperadorController, MaquinaController: verificar userId en PUT/DELETE para prevenir IDOR
- HoraTrabajada: cambiar horometroInicio de int a double para consistencia con horometroFin
- AuthController: email del remitente movido a variable de entorno BREVO_SENDER_EMAIL
- application.properties: agregar brevo.sender-email configurable

+++ src/main/java/com/maquicontrol/backend/model/HoraTrabajada.java
@@ -19,7 +19,7 @@ public class HoraTrabajada {
     private double horas;
     private String horaEntrada;
     private String horaSalida;
-    private int horometroInicio;
+    private double horometroInicio;
     private double horometroFin;
     private double valorHora;
     private LocalDate fecha;
@@ -53,8 +53,8 @@ public class HoraTrabajada {
     public String getHoraSalida() { return horaSalida; }
     public void setHoraSalida(String horaSalida) { this.horaSalida = horaSalida; }
 
-    public int getHorometroInicio() { return horometroInicio; }
-    public void setHorometroInicio(int horometroInicio) { this.horometroInicio = horometroInicio; }
+    public double getHorometroInicio() { return horometroInicio; }
+    public void setHorometroInicio(double horometroInicio) { this.horometroInicio = horometroInicio; }
 
     public double getHorometroFin() { return horometroFin; }
     public void setHorometroFin(double horometroFin) { this.horometroFin = horometroFin; }
